The \(R_{NX}(K)\) criterion

A curve indicating the improvement of the embedding over a random embedding for the neighborhood size \(K\). Values range from 0, for a random embedding, to 1 for a perfect embedding.

Usage

R_NX(Q)

Arguments

Q

a co-ranking matrix

Value

A vector with the values for R_NX(K)

Details

\(R_{NX}(K)\) is calculated as follows: $$ Q_{NX}(K) = \sum_{1\leq k\leq K}\sum_{1\leq l\leq K} \frac{q_{kl}}{KN} $$ Counts the upper left \(K\)-by-\(K\) block of \(Q\), i.e. it considers the preserved ranks on the diagonal and the permutations within a neighborhood.

$$ R_{NX}(K) = \frac{(N-1)Q_{NX}(K)-K}{N-1-K} $$ A resulting vale of 0 corresponds to a random embedding, a value of 1 to a perfect embedding of the \(K\)-ary neighborhood.

References

Lee, J.A., Lee, J.A., Verleysen, M., 2009. Quality assessment of dimensionality reduction: Rank-based criteria. Neurocomputing 72.

Lee, J. A., Peluffo-Ordóñez, D. H., & Verleysen, M., 2015. Multi-scale similarities in stochastic neighbour embedding: Reducing dimensionality while preserving both local and global structure. Neurocomputing, 169, 246–261. https://doi.org/10.1016/j.neucom.2014.12.095

Author

Guido Kraemer