Methods for the co-ranking matrix

coRanking provides methods for the calculation of the co-ranking matrix and derived measures to assess the quality of a dimensionality reduction

Details

This package provides functions for calculating the co-ranking matrix, plotitng functions and some derived measures for quality assessment of dimensionality reductions.

Funding provided by the Department for Biogeochemical Integration, Empirical Inference of the Earth System Group, at the Max Plack Institute for Biogeochemistry, Jena.

References

Chen, L., Buja, A., 2006. Local Multidimensional Scaling for Nonlinear Dimension Reduction, Graph Layout and Proximity Analysis.

Lee, J.A., Lee, J.A., Verleysen, M., 2009. Quality assessment of dimensionality reduction: Rank-based criteria. Neurocomputing 72.

Lueks, W., Mokbel, B., Biehl, M., & Hammer, B. (2011). How to Evaluate Dimensionality Reduction? - Improving the Co-ranking Matrix. ArXiv:1110.3917 [Cs]. http://arxiv.org/abs/1110.3917

Lee, J. A., Peluffo-Ordóñez, D. H., & Verleysen, M., 2015. Multi-scale similarities in stochastic neighbour embedding: Reducing dimensionality while preserving both local and global structure. Neurocomputing, 169, 246–261. https://doi.org/10.1016/j.neucom.2014.12.095

See also

Author

Maintainer: Guido Kraemer guido.kraemer@uni-leipzig.de (ORCID)